Added error handling for datagridview selection when casting selection to objects.
Removed GetById methods to be more compliant with OOP.

Solved bug when updating appointment with new doctor.

Added functionality to be able to sign out from current user and sign in with a new during the same compiler session.

Added some aesthetics so manage patients and appointments have similar layout.